requirements.txt
PyQt6
watchdog
Pillow
send2trash
